requirements.txt
blinker==1.9.0
click==8.2.1
colorama==0.4.6
Flask==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==6.0.0
MarkupSafe==3.0.2
pillow==11.3.0
python-docx==1.2.0
typing_extensions==4.14.1
Werkzeug==3.1.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
PyMySQL==1.1.0
email-validator==2.1.0
cryptography==42.0.5
Flask-Mail==0.10.0
PySide6
pyinstaller
pyserial==3.5
